Better support for table_row.find_index_by_title when tables have theads and tbodys.
In Watir, if the row we're looking at is in a tbody, we need to go up one level before finding the first row in the table.
For Selenium, if the row we're looking at is in a thead, accessing the 0th element of the thead gets nil. We have to go up one level to the table so [0] will return a row.