Add support for javascript alerts/confirms that reload page.
The existing functionality overrode window.alert and window.confirm in
order to save the alert/confirm message on the window object. However,
if the alert/confirm did a location.reload(), the message would be lost.
Instead of overriding these functions, use the built in alert/confirm
functionality of Watir and Selenium to save the alert message in a ruby
variable.