moving initialization of table rows and cells to main class
I've added and initialize row method to the table, and an initialize
Cell method to the table_row.  This simplifies extension of the table
class by pulling initialization of rows and cells out of the platform
modules, allowing extension of the Row or Cell class to be used more
easily.