Fix bug with selecting an option by text when using PageObject::PagePopulator#populate_page_with
The conditional to check whether the option passed was text would never
trigger as it was checking for the text against an array of
PageObject::Elements::Options rather than their text. Updated the
conditional to map the text of the options first.

Fix small typo in comment and yarddoc in page_populator.rb

Add feature test for populate_page_with functionality